Clone an Undirected Cyclic Graph
I had to read about the topic graph before doing it.
Expected Time Complexity: O(n + m).
Expected Auxiliary Space: O(n).

Given a connected undirected graph with n nodes and m edges, with each node having a distinct label from 0 to n-1, create a clone of the graph. Each node in the graph contains an integer val and an array (neighbors) of nodes, containing nodes that are adjacent to the current node.

Note: If the user returns a correct copy of the given graph, then the system will print 1; in the case when an incorrect copy is generated or when the user returns the original node, the system will print 0.