Max Index diff j-i such that A[i]<=A[j] & i<=j
Given an array a of n positive integers. The task is to find the maximum of j - i subjected to the constraint of a[i] < a[j] and i < j.
>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
n = 9
a[] = {34, 8, 10, 3, 2, 80, 30, 33, 1}
Output:
6
Explanation:
In the given array a[1] < a[7] satisfying the required condition(a[i] < a[j]) thus giving the maximum difference of j - i which is 6(7-1).